pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "specsmith"
version = "0.2.0"
description = "Forge governed project scaffolds from the Agentic AI Development Workflow Specification."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "BitConcepts"},
]
keywords = ["agentic", "scaffold", "governance", "agents-md", "cli", "specification"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.0",
    "types-pyyaml>=6.0",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]

[project.scripts]
specsmith = "specsmith.cli:main"

[project.urls]
Homepage = "https://github.com/BitConcepts/specsmith"
Documentation = "https://specsmith.readthedocs.io"
Repository = "https://github.com/BitConcepts/specsmith"
Changelog = "https://github.com/BitConcepts/specsmith/blob/main/CHANGELOG.md"
Issues = "https://github.com/BitConcepts/specsmith/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
specsmith = ["templates/**/*.j2"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
